pyproject.toml
[tool.poetry]
name = "opinionated-django"
version = "0.1.0"
description = ""
authors = ["Juan Olvera <juan@jolvera.dev>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
django-allauth = "^0.42.0"
django-anymail = {extras = ["mailgun"], version = "^8.0"}
django-storages = "^1.10.1"
django-environ = "^0.4.5"
django-on-heroku = "^1.0.1"
celery = "4.4.7"
django-celery-results = "^1.2.1"
django-celery-beat = "^2.0.0"
django-vote = "^2.2.0"
pillow = "^7.2.0"
boto3 = "^1.15.6"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
